data-pipeline/pyproject.toml
[project]
name = "data-pipeline"
version = "0.1.0"
description = "Notion KB indexer for support-cli"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "sentence-transformers>=2.7,<3.0",
    "opensearch-py>=2.4,<3.0",
    "python-dotenv>=1.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "mcp>=1.26.0",
    "boto3>=1.34",
]

[dependency-groups]
dev = [
    "debugpy>=1.8.20",
]
